Simplified new user stack creation
